BUGFIX: calculate reduced indices with country_name
The collapsed indices used to match the latent exec_constraint factor
to the regression was previously calculated grouping by country_name,
which meant that each group shared 1:N as the index values thereby
screwing up the indexing for the final onset regression. Instead, add
country_name to the pasted string and don't group anymore.

Definitely not the worlds most efficient or beautiful code, but it
should work.