Focus on all episodes, high intensity, and cumulative intensity
This is far from comprehensive for all the different coding standards
that can be applied to civil conflict onset, but focuses on the "big
three" which has been used by much of the previous
literature (ex. Bazzi & Blattman (2014), although they also
incorporate a number of datasets besides just UCDP/Prio).