BUGFIX: Properly check bitarray buffer capacity and resize
First bug is an off by one error when checking the current capacity of
the buffer for a bitarray. The second bug concerns the memset after
resizing being only applied to the first N bytes, rather than the
newly allocated area.

Neither affect the presented analysis since the bitarray to track
matching columns was large enough to never be resized.